config reorgs
